+++ src/Main.java
@@ -2,7 +2,7 @@ public class Main {
     public static void main(String[] args) {
         BmiService service = new BmiService();
 
-        double index = service.calculate(55, 1.67);
+        int index = service.calculate(55, 1.67);
 
         System.out.println(index);
     }
